requirements.txt
flask
flask-cors
flask-socketio
pymongo
python-dotenv
reportlab
pandas
requests
opencv-python-headless
numpy
Pillow
flask
flask-cors
flask-socketio
flask-jwt-extended
flask-babel
flask-login
pymongo
requests
numpy
opencv-python-headless
scikit-learn
reportlab
pandas
python-dotenv
apscheduler
geopy
gunicorn
Flask==2.3.2
Flask-Cors==3.0.10
Flask-SocketIO==5.3.4
python-dotenv==1.0.0
pymongo==4.3.3
eventlet==0.33.3
dnspython==2.5.0
